requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
pymupdf
python-dotenv
